Updated the hardware schematics.
Fixed the main problem that was causing the BME280 sensor to seize I2C
bus transmissions. The problem was that transient RF interference on
the CSB line was able to switch the device to the SPI communication
mode, and it was impossible to switch it back by means other than
performing a power-on reset for the sensor. From the BME280 datasheet,
it was unclear if such CSB line behavior was taking place only during
device startup or perpetually. Tying the CSB line to VCC directly
should improve device stability as the built-in 10k pull-up resistor
on the line appears not sufficient.